* Create mode_test.go * Move test for mode detection to mode_test.go * Fix GetInstructionMode for absolute and zeropage
